Update swinging door and add sliding door component
Update the swinging door to allow for swinging both ways, and allow the door to close if the player isn't near it for long enough.

Add a sliding door component with an Open function that is exposed through blueprints.

Update the maze level to require two separate button presses to open a big sliding door.